Cleanup runtime includes, separate types.cpp and utils.cpp, add testing library that includes testing utils and emitc dylib APIs isolated under runtime/test/ttnn
