+++ runtime/test/ttnn/dylib.cpp
@@ -2,11 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "tt/runtime/test/dylib.h"
+#include "tt/runtime/ttnn/test/dylib.h"
 
 #include "tt/runtime/detail/logger.h"
 #include "tt/runtime/runtime.h"
-#include "tt/runtime/test/utils.h"
 #include "tt/runtime/ttnn/types.h"
 #include "tt/runtime/ttnn/utils.h"
 #include "tt/runtime/types.h"
