+++ runtime/test/ttnn/gtest/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_runtime_gtest(subtract_test test_subtract.cpp)
